+++ cmd/config.go
@@ -1,0 +1,15 @@
+package cmd
+
+import (
+	"github.com/spf13/cobra"
+)
+
+// configCmd represents the config command
+var configCmd = &cobra.Command{
+	Use:   "config",
+	Short: "Manage configs",
+}
+
+func init() {
+	rootCmd.AddCommand(configCmd)
+}